refactor: migrate WeniIntegrationsClient to integrations module and update references
- Moved WeniIntegrationsClient from insights.sources.wabas.clients to insights.sources.integrations.clients for better organization.
- Updated all references in the codebase to use the new client location, ensuring consistent functionality across services and tests.
- Adjusted method signatures to align with the new client initialization.
