+++ insights/metrics/meta/permissions.py
@@ -1,6 +1,6 @@
 from rest_framework.permissions import BasePermission
 
-from insights.sources.wabas.clients import WeniIntegrationsClient
+from insights.sources.integrations.clients import WeniIntegrationsClient
 
 
 class ProjectWABAPermission(BasePermission):
@@ -18,8 +18,8 @@ class ProjectWABAPermission(BasePermission):
         if not project_uuid or not waba_id:
             return False
 
-        integrations_client = WeniIntegrationsClient(project_uuid=project_uuid)
-        wabas_data = integrations_client.get_wabas_for_project()
+        integrations_client = WeniIntegrationsClient()
+        wabas_data = integrations_client.get_wabas_for_project(project_uuid)
         wabas_ids = {waba.get("waba_id") for waba in wabas_data}
 
         return waba_id in wabas_ids
